add new motd feature
