+++ catris/high_scores.py
@@ -5,29 +5,14 @@ import dataclasses
 import io
 import logging
 import re
-import sys
 from typing import IO, TYPE_CHECKING, Iterator
 
+from catris.compat import to_thread
 from catris.games import Game
 from catris.views import GameOverView, PlayingView
 
 if TYPE_CHECKING:
     from catris.lobby import Lobby
-
-if sys.version_info >= (3, 9):
-    from asyncio import to_thread
-else:
-    from typing import Any
-
-    # copied from source code with slight modifications
-    async def to_thread(func: Any, *args: Any, **kwargs: Any) -> Any:
-        import contextvars
-        import functools
-
-        loop = asyncio.get_running_loop()
-        ctx = contextvars.copy_context()
-        func_call = functools.partial(ctx.run, func, *args, **kwargs)
-        return await loop.run_in_executor(None, func_call)
 
 
 _logger = logging.getLogger(__name__)
